# Optimisation des requêtes SQL sur la page d'accueil