… zds/tutorialv2/managers.py