Optimisation des requêtes SQL sur la page d'accueil

**Avant**
- Utilisateur connecté : 132 requêtes pour la première visite et 101 pour les suivantes (~2s)
- Anonyme : 103 requêtes pour la première visite et 74 pour les suivantes (~1,6s)

**Après**
- Utilisateur connecté : 100 requêtes pour la première visite et 69 pour les suivantes (~1,7s)
- Anonyme : 79 requêtes pour la première visite et 50 pour les suivantes (~1,4s)

La différence entre la première requête et les autres est due au cache.